+++ src/hooks/useModal.ts
@@ -10,6 +10,7 @@ const useModal = () => {
   };
   return {
     isModalOpen,
+    setIsModalOpen,
     handleShowModal,
     handleCloseModal,
   };
